Add checkbox to search only directory names
When the checkbox is checked, only directories are matched against
"Filename" field. Files are not matched and no file is searched for
content even if the field "Content" is not empty.

This is an example feature implemented for open source programming course
at Czech Technical University. It is intentionally based on older version,
so that students can learn how to use git to resolve merge conflicts or
update (rebase) their patches.

Signed-off-by: Michal Sojka <sojkam1@fel.cvut.cz>